+++ config/husky.js
@@ -1,6 +1,0 @@
-module.exports = {
-    hooks: {
-        'pre-commit':
-            'lint-staged -c @creativestyle/magesuite-frontend-builder/config/lint-staged',
-    },
-};