+++ gulp/config/clean.js
@@ -1,17 +1,13 @@
-const path = require('path');
+import path from 'path';
 
-const paths = require('../paths');
+import paths from '../paths.js';
 
 /**
  * Cleaning task settings.
  */
-module.exports = {
+export default {
     /**
      * Paths that should be deleted.
      */
-    src: [
-        path.join(paths.dist, '**/*'),
-        path.resolve('.cache-loader'),
-        path.resolve('node_modules/.cache/cache-loader'),
-    ],
+    src: [path.join(paths.dist, '**/*')],
 };