+++ gulp/config/collectViewXml.js
@@ -1,11 +1,12 @@
-const path = require('path');
-const paths = require('../paths');
-const parentAliases = require('../parentAliases');
+import path from 'path';
+
+import paths from '../paths.js';
+import parentAliases from '../parentAliases.js';
 
 /**
  * Configuration for collecting view.xml files.
  */
-module.exports = {
+export default {
     src: [
         path.resolve('../../magento/theme-frontend-blank'),
         path.resolve('../magesuite-content-constructor-frontend'),