+++ gulp/config/copyDocs.js
@@ -1,11 +1,12 @@
 // @ts-check
-const path = require('path');
-const paths = require('../paths');
+import path from 'path';
+
+import paths from '../paths.js';
 
 /**
  *  Configuration for documentation copy task.
  */
-module.exports = {
+export default {
     watch: [
         // Images except sprites
         path.join(paths.src, 'doc/**/*.{md,gif,png,jpg,webp,svg,ico}'),